Added Custom Log Appender and refactored Log Excerpt into UserControl
Added a custom logging appender that keeps an excerpt of recent logging events.  This output is bound to the Log view in the UI, that I've refactored into a UserControl with dependency properties.  This project should be mostly complete now.  The only thing left on the to-do list is possibly a Mock view model for the Designer.